Resolves issues and splits large House Hearing uploads into separate batches.

+++ src/main/java/edu/temple/cla/policydb/ppdpapp/api/tables/LegServAgncyReports.java
@@ -39,6 +39,7 @@ import edu.temple.cla.policydb.ppdpapp.api.models.Batch;
 import edu.temple.cla.policydb.ppdpapp.api.models.File;
 import edu.temple.cla.policydb.ppdpapp.api.models.MetaData;
 import edu.temple.cla.policydb.ppdpapp.util.ZipUtil;
+import static edu.temple.cla.policydb.ppdpapp.util.ZipUtil.isZipFile;
 import java.io.BufferedOutputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -56,7 +57,6 @@ import java.util.Objects;
 import java.util.StringJoiner;
 import static java.util.stream.Collectors.toList;
 import javax.persistence.Tuple;
-import org.apache.log4j.Logger;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
@@ -129,9 +129,11 @@ public class LegServAgncyReports extends AbstractTable {
                 URL url = new URL(hyperlink);
                 String fullPathName = url.toURI().getPath();
                 java.io.File sourceFile = new java.io.File(fullPathName);
-                String fileName = sourceFile.getName();
-                java.io.File javaFile = enterFileIntoDatabase(docObj, fileName);
-                sourceFile.renameTo(javaFile);
+                if (sourceFile.exists() && !isZipFile(sourceFile.getName())) {
+                    String fileName = sourceFile.getName();
+                    java.io.File javaFile = enterFileIntoDatabase(docObj, fileName);
+                    sourceFile.renameTo(javaFile);
+                }
             } catch (Exception e) {
                 throw new RuntimeException("Error entering LSAR file into database", e);
             }
@@ -354,6 +356,7 @@ public class LegServAgncyReports extends AbstractTable {
     /**
      * Method to decompress zip file and add the content documents to a batch.
      * This method is currently only applicable to LegServiceAgencyReports.
+     * If the file is not a zip file, this method returns the Batch object.
      * @param documentDAO The Document DAO
      * @param fileDAO The File DAO
      * @param batchDAO The Batch DAO
@@ -361,13 +364,14 @@ public class LegServAgncyReports extends AbstractTable {
      * @return Updated Batch object, or error indication.
      */
     @Override
-    public ResponseEntity<?> checkZip(DocumentDAO documentDAO, FileDAO fileDAO, 
+    public ResponseEntity<?> addToBatch(DocumentDAO documentDAO, FileDAO fileDAO, 
             BatchDAO batchDAO, Batch batchObj) {
         if (batchObj.getFileID() != null) {
             int fileId;
             try {
                 fileId = Integer.parseInt(batchObj.getFileID());
             } catch (NumberFormatException ex) {
+                // Batch not associated with a file
                 return new ResponseEntity<>(batchObj, HttpStatus.OK);
             }
             File file = fileDAO.find(fileId);
@@ -380,9 +384,15 @@ public class LegServAgncyReports extends AbstractTable {
                     int docID = documentDAO.insertDocument(getTableName(), lsaReportObject);
                     batchDAO.addDocument(batchObj.getBatchID(), Integer.toString(docID));
                 }
-            }
+                return new ResponseEntity<>(batchObj, HttpStatus.OK);            
+            } else {
+                // File is not a ZIP file.
+                return new ResponseEntity<>(batchObj, HttpStatus.OK);
+            }            
+        } else {
+            // Batch not associated with a file.
+            return new ResponseEntity<>(batchObj, HttpStatus.OK);
         }
-        return new ResponseEntity<>(batchObj, HttpStatus.OK);
     }
     
     private Map<String, Object> createLSAReport(java.io.File javaFile) {
